Suppresses extra empty lines in output SII files.
Extra blank lines were being added to the torque curve when editing
engines that used the default torque curve. This was because the regex
was capturing the line break while the default torque code also had
line breaks, thus double line breaks.

In order to improve readability in the code, the regex capture group
now exempts the line break character in the text, while it still captures
the line break for engines that already have torque curves.